API/requirements.txt
fastapi==0.65.1
sqlalchemy==1.4.18
python-jenkins==1.7.0
uvloop==0.15.2
httptools==0.1.2
psycopg2-binary==2.9
python-multipart==0.0.5
uvicorn==0.15.0
cerberus==1.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyYAML==6.0
ftputil==5.0.4